Create log management for Errorlog and slowquery and fix duplicate backup jobs due to waiting socat logs

+++ share/scripts/dbjobs_new.sh
@@ -25,22 +25,6 @@ MYSQL_CLIENT="%%ENV:SVC_CONF_ENV_CLIENT_BASEDIR%%/mysql"
 MYSQL_CHECK=%%ENV:SVC_CONF_ENV_CLIENT_BASEDIR%%/mysqlcheck
 MYSQL_DUMP=%%ENV:SVC_CONF_ENV_CLIENT_BASEDIR%%/mysqldump
 
-# Determine which binary to use (prefer MariaDB, fallback to MySQL)
-if [ -x "$MARIADB_CLIENT" ]; then
-    BINARY_CLIENT="$MARIADB_CLIENT $BINARY_CLIENT_PARAMETERS"
-    BINARY_CHECK=$MARIADB_CHECK
-    BINARY_DUMP=$MARIADB_DUMP
-    echo "Job start: Using MariaDB binaries."
-elif [ -x "$MYSQL_CLIENT" ]; then
-    BINARY_CLIENT="$MYSQL_CLIENT $BINARY_CLIENT_PARAMETERS"
-    BINARY_CHECK=$MYSQL_CHECK
-    BINARY_DUMP=$MYSQL_DUMP
-    echo "Job start: Using MySQL binaries."
-else
-    echo "Neither MariaDB nor MySQL binaries are available."
-    exit 1
-fi
-
 SST_RECEIVER_PORT=%%ENV:SVC_CONF_ENV_SST_RECEIVER_PORT%%
 SOCAT_BIND=%%ENV:SERVER_IP%%
 MARIADB_BACKUP=%%ENV:SVC_CONF_ENV_CLIENT_BASEDIR%%/mariabackup
@@ -61,34 +45,18 @@ LOCK_DIR="$TMP_DIR/locks"
 BATCH_SIZE=5
 JOBS=("xtrabackup" "mariabackup" "errorlog" "slowquery" "zfssnapback" "optimize" "reseedxtrabackup" "reseedmariabackup" "flashbackxtrabackup" "flashbackmariadbackup" "stop" "restart" "start")
 
-table_exists() {
-    local db="$1"
-    local table="$2"
-
-    # Ask information_schema if the table exists
-    local result
-    result=$(echo "SELECT COUNT(*) 
-                   FROM information_schema.tables 
-                   WHERE table_schema='${db}' 
-                     AND table_name='${table}';" \
-             | $BINARY_CLIENT -N 2>/dev/null)
-
-    if [ "$result" = "1" ]; then
-        return 0  # true, table exists
-    else
-        return 1  # false, table missing
-    fi
-}
-
-create_jobs_table() {
-    if ! table_exists "replication_manager_schema" "jobs"; then
-        echo "Creating jobs table..."
-        echo "set sql_log_bin=0;CREATE DATABASE IF NOT EXISTS replication_manager_schema;CREATE TABLE IF NOT EXISTS replication_manager_schema.jobs(id INT NOT NULL auto_increment PRIMARY KEY, task VARCHAR(20),  port INT, server VARCHAR(255), done TINYINT not null default 0, state tinyint not null default 0, result MEDIUMTEXT, start DATETIME, end DATETIME, KEY idx1(task,done) ,KEY idx2(result(1),task), KEY idx3 (task, state), UNIQUE(task)) engine=innodb;set sql_log_bin=1;" | $BINARY_CLIENT
-    fi
-}
-
 # OSX need socat extra path
 export PATH=$PATH:/usr/local/bin
+
+# Logging levels
+LVL_ERROR="ERROR"
+LVL_WARN="WARN"
+LVL_INFO="INFO"
+LVL_DEBUG="DEBUG"
+
+########################
+# Function Definitions #
+########################
 
 pad_pkcs7() {
     local data="$1"
@@ -170,45 +138,116 @@ send_encrypted_data() {
 send_lines_to_api() {
     local lines="$1"
     local job="$2"
+    local level="$3"
     local address="${REPLICATION_MANAGER_URL}"
-    local data="{\"server\":\"$MYSQL_SERVER:$MYSQL_PORT\",\"log\":\"$lines\"}"
-
     local max_retries=3
     local attempt=0
     local success=false
 
+    if [ -z "$job" ]; then
+        job="main"
+        level="$LVL_DEBUG"
+    fi
+
+    if [ -z "$level" ]; then
+        level="$LVL_DEBUG"
+    fi
+
+    local data="{\"server\":\"$MYSQL_SERVER:$MYSQL_PORT\",\"log\":\"$lines\",\"level\":\"$level\"}"
+    local http_code
+    local response
+
     while ((attempt < max_retries)); do
         # Capture response and HTTP status code
-        local response
         response=$(send_encrypted_data "$address" "$job" "$data")
 
-        # echo "$response" >> $LOG_DIR/curl_response.txt
-        
         # Extract HTTP status code
-        local http_code=$(echo "$response" | grep -oP '(?<=HTTP/1.1 )[0-9]{3}')
+        http_code=$(echo "$response" | grep -oP '(?<=HTTP/1.1 )[0-9]{3}')
 
         if [[ "$http_code" == "200" ]]; then
-            echo "API call successful for job: $job"
+            # echo "API call successful for job: $job"
             success=true
             break
         else
-            echo "API call failed for job: $job with status code: $http_code"
-            cat $LOG_DIR/curl_response.txt
+            # echo "API call failed for job: $job with status code: $http_code"
+            # cat $LOG_DIR/curl_response.txt
             ((attempt++))
             sleep 2 # Wait before retrying
         fi
     done
 
     if [ "$success" = false ]; then
-        echo "API call failed after $max_retries attempts for job: $job"
+        # Backup log file if it exceeds 1MB. Only keep1 old log file.
+        if [ -f "$LOG_DIR/api_calls.log" ]; then
+            local filesize
+            filesize=$(stat -c%s "$LOG_DIR/api_calls.log")
+            if ((filesize > 1048576)); then
+                cp -f "$LOG_DIR/api_calls.log" "$LOG_DIR/api_calls.log.bak"
+                : > "$LOG_DIR/api_calls.log"
+            fi
+        else 
+            touch "$LOG_DIR/api_calls.log"
+        fi
+
+        echo "API call failed with code: $http_code after $max_retries attempts for job: $job" >> $LOG_DIR/api_calls.log
+        echo "Request Destination: $address Job: $job Data: $data" >> $LOG_DIR/api_calls.log
+        echo "$response" >> $LOG_DIR/api_calls.log
     fi
 }
+
+# Determine which binary to use (prefer MariaDB, fallback to MySQL)
+if [ -x "$MARIADB_CLIENT" ]; then
+    BINARY_CLIENT="$MARIADB_CLIENT $BINARY_CLIENT_PARAMETERS"
+    BINARY_CHECK=$MARIADB_CHECK
+    BINARY_DUMP=$MARIADB_DUMP
+    send_lines_to_api "Job start: Using MariaDB binaries."
+elif [ -x "$MYSQL_CLIENT" ]; then
+    BINARY_CLIENT="$MYSQL_CLIENT $BINARY_CLIENT_PARAMETERS"
+    BINARY_CHECK=$MYSQL_CHECK
+    BINARY_DUMP=$MYSQL_DUMP
+    send_lines_to_api "Job start: Using MySQL binaries."
+else
+    send_lines_to_api "Neither MariaDB nor MySQL binaries are available. Exiting job script." "main" "$LVL_ERROR"
+    exit 1
+fi
+
+# Function to check if a table exists in a given database
+table_exists() {
+    local db="$1"
+    local table="$2"
+
+    # Ask information_schema if the table exists
+    local result
+    result=$(echo "SELECT COUNT(*) 
+                   FROM information_schema.tables 
+                   WHERE table_schema='${db}' 
+                     AND table_name='${table}';" \
+             | $BINARY_CLIENT -N 2>/dev/null)
+
+    if [ "$result" = "1" ]; then
+        return 0  # true, table exists
+    else
+        return 1  # false, table missing
+    fi
+}
+
+# Function to create the jobs table if it doesn't exist
+create_jobs_table() {
+    if ! table_exists "replication_manager_schema" "jobs"; then
+        send_lines_to_api "Creating jobs table..." "main" "$LVL_INFO"
+        echo "set sql_log_bin=0;CREATE DATABASE IF NOT EXISTS replication_manager_schema;CREATE TABLE IF NOT EXISTS replication_manager_schema.jobs(id INT NOT NULL auto_increment PRIMARY KEY, task VARCHAR(20),  port INT, server VARCHAR(255), done TINYINT not null default 0, state tinyint not null default 0, result MEDIUMTEXT, start DATETIME, end DATETIME, KEY idx1(task,done) ,KEY idx2(result(1),task), KEY idx3 (task, state), UNIQUE(task)) engine=innodb;set sql_log_bin=1;" | $BINARY_CLIENT
+    fi
+}
+
+################################
+# Log Processing Functions     #
+################################
 
 # Function to create a manual lock file
 create_lock_file() {
     local lock_file="$1"
     if [ -e "$lock_file" ]; then
-        echo "Lock file exists. Exiting."
+        send_lines_to_api "Lock file $lock_file exists. Exiting." "main" "$LVL_ERROR"
         return 1
     fi
     touch "$lock_file"
@@ -218,7 +257,17 @@ create_lock_file() {
 # Function to remove a manual lock file
 remove_lock_file() {
     local lock_file="$1"
-    rm -f "$lock_file"
+    if [ -e "$lock_file" ]; then
+        rm -f "$lock_file"
+    fi
+}
+
+# Function to remove a run directory lock file
+remove_run_lockdir() {
+    local run_lockdir="$1"
+    if [ -d "$run_lockdir" ]; then
+        rmdir "$run_lockdir"
+    fi
 }
 
 # Function to wait for the .run file with a timeout
@@ -227,17 +276,17 @@ wait_for_run_lockdir() {
     local timeout=30
     local start_time=$(date +%s)
 
-    send_lines_to_api "Waiting for $run_lockdir file...\n" "$job"
+    send_lines_to_api "Waiting for $run_lockdir file...\n" "$job" "$LVL_DEBUG"
     while [[ ! -d "$run_lockdir" ]]; do
         sleep 0.5
         local current_time=$(date +%s)
         local elapsed=$((current_time - start_time))
         if ((elapsed >= timeout)); then
-            send_lines_to_api "Timeout reached while waiting for .run file.\n" "$job"
+            send_lines_to_api "Timeout reached while waiting for .run file.\n" "$job" "$LVL_ERROR"
             return 1
         fi
     done
-    send_lines_to_api "$run_lockdir file found...\n" "$job"
+    send_lines_to_api "$run_lockdir file found...\n" "$job" "$LVL_DEBUG"
     return 0
 }
 
@@ -247,17 +296,17 @@ wait_for_log_file() {
     local timeout=60
     local start_time=$(date +%s)
 
-    send_lines_to_api "Waiting for $logfile file...\n" "$job"
+    send_lines_to_api "Waiting for $logfile file...\n" "$job" "$LVL_DEBUG"
     while [[ ! -f "$logfile" ]]; do
         sleep 0.5
         local current_time=$(date +%s)
         local elapsed=$((current_time - start_time))
         if ((elapsed >= timeout)); then
-            send_lines_to_api "Timeout reached while waiting for $logfile file. Please check log manually if needed. \n" "$job"
+            send_lines_to_api "Timeout reached while waiting for $logfile file. Please check log manually if needed. \n" "$job" "$LVL_ERROR"
             return 1
         fi
     done
-    send_lines_to_api "$logfile file found...\n" "$job"
+    send_lines_to_api "$logfile file found...\n" "$job" "$LVL_DEBUG"
     return 0
 }
 
@@ -279,13 +328,13 @@ read_log_file() {
             ((current_line++))
 
             if [[ ! -d "$run_lockdir" ]]; then
-                send_lines_to_api "Run file has been deleted. Processing remaining lines.\n" "$job"
+                send_lines_to_api "Run file has been deleted. Processing remaining lines.\n" "$job" "$LVL_DEBUG"
                 break
             fi
 
             batch+="$escaped\n"
             if ((current_line % BATCH_SIZE == 0)); then
-                send_lines_to_api "$batch" "$job"
+                send_lines_to_api "$batch" "$job" "$LVL_DEBUG"
                 batch=""
             fi
             echo "$current_line" >"$checkpoint_file"
@@ -294,7 +343,7 @@ read_log_file() {
 
         # Send any remaining lines in the batch after the first loop
         if [[ -n "$batch" ]]; then
-            send_lines_to_api "$batch" "$job"
+            send_lines_to_api "$batch" "$job" "$LVL_DEBUG"
         fi
     fi
 }
@@ -326,7 +375,7 @@ process_log_file() {
         return
     fi
 
-    # Ensure lock file is removed on script exit
+    # Ensure lock file is removed on script exit. Using trap to handle unexpected exit. This will not interfere with other traps since it's a separate subshell.
     trap 'remove_lock_file "$lock_file"' EXIT
 
     if ! wait_for_run_lockdir "$run_lockdir"; then
@@ -344,7 +393,7 @@ process_log_file() {
         last_line=$(cat "$checkpoint_file")
     fi
 
-    send_lines_to_api "Last checkpoint on "$checkpoint_file" is: $last_line.\n" "$job"
+    send_lines_to_api "Last checkpoint on "$checkpoint_file" is: $last_line.\n" "$job" "$LVL_DEBUG"
 
     local current_line=0
     local batch=""
@@ -363,7 +412,7 @@ process_log_file() {
             ((current_line++))
             batch+="$escaped\n"
             if ((current_line % BATCH_SIZE == 0)); then
-                send_lines_to_api "$batch" "$job"
+                send_lines_to_api "$batch" "$job" "$LVL_DEBUG"
                 batch=""
             fi
             echo "$current_line" >"$checkpoint_file"
@@ -371,14 +420,18 @@ process_log_file() {
     fi
 
     if [[ -n "$batch" ]]; then
-        send_lines_to_api "$batch" "$job"
+        send_lines_to_api "$batch" "$job" "$LVL_DEBUG"
     fi
 
-    send_lines_to_api "Removing checkpoint file.\n" "$job"
+    send_lines_to_api "Removing checkpoint file.\n" "$job" "$LVL_DEBUG"
     rm -f "$checkpoint_file"
 
     remove_lock_file "$lock_file"
 }
+
+#################################
+# Job Related Functions     #
+#################################
 
 socatCleaner() {
     kill -9 $(lsof -t -i:$SST_RECEIVER_PORT -sTCP:LISTEN)
@@ -415,9 +468,9 @@ doneJob() {
     esac
     
     if [ $jobstate -eq 3 ]; then
-        send_lines_to_api "Job $job ended with state: Finished" "$job" 
+        send_lines_to_api "Job $job ended with state: Finished" "$job" "$LVL_INFO"
     else
-        send_lines_to_api "Job $job ended with state: Error" "$job"
+        send_lines_to_api "Job $job ended with state: Error" "$job" "$LVL_ERROR"
     fi
     $BINARY_CLIENT -e "set sql_log_bin=0;UPDATE replication_manager_schema.jobs set end=NOW(), state=$jobstate, result=LOAD_FILE('$LOG_DIR/$job.out'), done=$done  WHERE id='$ID';" &
 }
@@ -427,11 +480,11 @@ pauseJob() {
 }
 
 partialRestore() {
-    send_lines_to_api "Starting partial restore..." "$job" 
+    send_lines_to_api "Starting partial restore..." "$job" "$LVL_INFO"
     chown -R mysql:mysql $BACKUPDIR 
     $BINARY_CLIENT -e "set sql_log_bin=0;install plugin BLACKHOLE soname 'ha_blackhole.so'"
     for dir in $(ls -d $BACKUPDIR/*/ | xargs -n 1 basename | grep -vE 'mysql|performance_schema|replication_manager_schema'); do
-        send_lines_to_api "Restoring $dir..." "$job" 
+        send_lines_to_api "Restoring $dir..." "$job" "$LVL_DEBUG"
         $BINARY_CLIENT -e "set sql_log_bin=0;drop database IF EXISTS $dir; CREATE DATABASE $dir;"
 
         for file in $(find $BACKUPDIR/$dir/ -name "*.ibd" | xargs -n 1 basename | cut -d'.' --complement -f2-); do
@@ -457,10 +510,11 @@ partialRestore() {
         mv $BACKUPDIR/mysql/$file.* $DATADIR/mysql/
         $BINARY_CLIENT -e "set sql_log_bin=0;FLUSH TABLE mysql.$file"
     done
-    send_lines_to_api "Setting GTID of the last change..." "$job" 
+    send_lines_to_api "Setting GTID of the last change..." "$job" "$LVL_DEBUG"
     cat $BACKUPDIR/xtrabackup_info | grep binlog_pos | awk -F, '{ print $3 }' | sed -e 's/GTID of the last change/set sql_log_bin=0;set global gtid_slave_pos=/g' | $BINARY_CLIENT
-    send_lines_to_api "Flushing privileges..." "$job" 
+    send_lines_to_api "Flushing privileges..." "$job" "$LVL_DEBUG"
     $BINARY_CLIENT -e"set sql_log_bin=0;flush privileges;start slave;"
+    send_lines_to_api "Partial restore done." "$job" "$LVL_INFO"
 }
 
 #######################
@@ -502,7 +556,7 @@ for job in "${JOBS[@]}"; do
     ID=($(echo $TASK | awk -F@ '{ print $1 }'))
 
     if [ "$ID" != "" ]; then
-        send_lines_to_api "Job $job initiated. Clearing previous logs..." "$job" 
+        send_lines_to_api "Job $job initiated. Clearing previous logs..." "$job" "$LVL_INFO"
         case "$job" in
             mariabackup|xtrabackup)
                 rm -f "$LOG_DIR/backup.out"
@@ -534,7 +588,8 @@ for job in "${JOBS[@]}"; do
 
         mkdir -p "$LOG_DIR/$job.run"
         process_log_file "$job" &
-        trap 'rmdir "$LOG_DIR/$job.run"' EXIT
+        # Ensure run lockdir for current job is removed on script exit. Intended to replace the previous job trap which already removed at the end of loop entry.
+        trap 'remove_run_lockdir "$LOG_DIR/$job.run"' EXIT
         echo "Processing $job"
         
         #purge de past
@@ -607,7 +662,7 @@ for job in "${JOBS[@]}"; do
                 fi
                 >$ERROLOG
             else 
-                echo -n | socat -u stdio TCP:$ADDRESS
+                echo -n | socat -u stdio TCP:$ADDRESS &>"$LOG_DIR/$job.process.out"
             fi
             ;;
         slowquery)
@@ -626,7 +681,7 @@ for job in "${JOBS[@]}"; do
                 fi
                 >$SLOWLOG
             else 
-                echo -n | socat -u stdio TCP:$ADDRESS
+                echo -n | socat -u stdio TCP:$ADDRESS &>"$LOG_DIR/$job.process.out"
             fi
             ;;
         zfssnapback)
@@ -648,6 +703,7 @@ for job in "${JOBS[@]}"; do
             ;;
         esac
         doneJob "$job"
-        sleep 1 && rmdir "$LOG_DIR/$job.run" &
+        sleep 1 && remove_run_lockdir "$LOG_DIR/$job.run" &
+        trap - EXIT
     fi
 done
